requirements.txt
Flask-SQLAlchemy==2.1
flask-restplus==0.9.2
marshmallow-sqlalchemy==0.14.1
flask-marshmallow==0.9.0
prometheus_client==0.6.0